requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
colorama==0.4.4
idna==2.10
numpy==1.19.2
pandas==1.1.3
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
shodan==1.24.0
six==1.15.0
urllib3==1.25.11
XlsxWriter==1.3.7
